fix bug

+++ src/components/editors/unityModelEditor.tsx
@@ -1,4 +1,4 @@
-import React, { ChangeEvent, useEffect, useRef, useState } from 'react'
+import React, { ChangeEvent, useCallback, useEffect, useRef, useState, FocusEvent } from 'react'
 import { SelectableValue, StandardEditorProps } from "@grafana/data";
 import { Field, Select, Input, FileDropzone, FileListItem } from '@grafana/ui';
 import { defaultUnityModel } from 'utils/defaults';
@@ -17,12 +17,20 @@ export const UnityModelEditor: React.FC<Props> = ({ item, value, onChange }) => 
     const modeOptions: ISelect[] = enumToSelect(SetUnityModelMode)
 
     const [mode, setMode] = useState<SelectableValue<SetUnityModelMode>>({ label: value.setFilesMode, value: value.setFilesMode })
+    const [unityFiles, setUnityFiles] = useState<UnityFiles>(defaultUnityModel.unityFiles)
 
-    const handleOnChangeInput = (event: ChangeEvent<HTMLInputElement>) => {
+    
+    const handleOnBlurInput = (event: FocusEvent<HTMLInputElement>) => {
         if (value.unityFiles) {
-            value.unityFiles[event.currentTarget.name as keyof UnityFiles] = event.target.value;
+            value.unityFiles[event.currentTarget.name as keyof UnityFiles] = event.target.value
             onChange(value)
         }
+    }
+
+    const handleOnChangeInput = (event: ChangeEvent<HTMLInputElement>) => {
+        let aux = {...unityFiles}
+        aux[event.currentTarget.name as keyof UnityFiles] = event.target.value
+        setUnityFiles(aux)
     }
 
     const handleOnFileUpload = (result: string | ArrayBuffer | null, key: string) => {
@@ -38,6 +46,7 @@ export const UnityModelEditor: React.FC<Props> = ({ item, value, onChange }) => 
     }
 
     const clearFields = () => {
+        setUnityFiles(defaultUnityModel.unityFiles)
         value.unityFiles = defaultUnityModel.unityFiles
         onChange(value)
     }
@@ -55,7 +64,7 @@ export const UnityModelEditor: React.FC<Props> = ({ item, value, onChange }) => 
     }
 
     const getFileItemIfAny = (name: NamesUnityModel, fileName: string) => {
-        const file = value.unityFiles[name as keyof UnityFiles]
+        const file = unityFiles[name as keyof UnityFiles]
         if (file !== undefined && file.length > 0) {
             //const blob: any = new Blob([file]); //
             return <FileListItem file={{ file: new File([], fileName), id: fileName, error: null }}
@@ -73,12 +82,12 @@ export const UnityModelEditor: React.FC<Props> = ({ item, value, onChange }) => 
         </Field>
     }
 
-    const InputComponent = ({ name, suffix }: { name: NamesUnityModel, suffix: string }) => {
-        let val: string = value.unityFiles[name as keyof UnityFiles]
+    const InputComponent = useCallback(({ value, name, suffix }: { value: string, name: NamesUnityModel, suffix: string }) => {
+        console.log(name, value)
         return <Field label={suffix}>
-            <Input name={name} value={val} required onChange={handleOnChangeInput} />
+            <Input name={name} value={value} required onChange={handleOnChangeInput} onBlur={handleOnBlurInput}/>
         </Field>
-    }
+    }, [value])
 
     const uploadFileMode = <div ref={divRef}>
         <FileDropComponent name="dataUrl" suffix=".data" />
@@ -88,10 +97,10 @@ export const UnityModelEditor: React.FC<Props> = ({ item, value, onChange }) => 
     </div>
 
     const setInputMode = <div>
-        <InputComponent name="dataUrl" suffix=".data" />
-        <InputComponent name="frameworkUrl" suffix=".framework.js" />
-        <InputComponent name="loaderUrl" suffix=".loader.js" />
-        <InputComponent name="codeUrl" suffix=".wasm" />
+        <InputComponent value={unityFiles.dataUrl} name="dataUrl" suffix=".data" />
+        <InputComponent value={unityFiles.frameworkUrl} name="frameworkUrl" suffix=".framework.js" />
+        <InputComponent value={unityFiles.loaderUrl} name="loaderUrl" suffix=".loader.js" />
+        <InputComponent value={unityFiles.codeUrl} name="codeUrl" suffix=".wasm" />
     </div>
 
     useEffect(() => {
@@ -103,10 +112,18 @@ export const UnityModelEditor: React.FC<Props> = ({ item, value, onChange }) => 
     }, [mode])
 
     useEffect(() => {
-        deleteSizeInfo()
+        if (mode.value === SetUnityModelMode.dragDrop) {
+            deleteSizeInfo()
+        }
     }, [getFileItemIfAny])
-    
 
+    useEffect(() => {
+        setUnityFiles(value.unityFiles)
+    }, [])
+
+    useEffect(() => {
+        console.log("AAAAAAAAAAAAAAAAAA")
+    }, [value])
 
     const getConfigByMode = () => {
         switch (mode.value) {
